Add support for Active Record 6.1.x.
A note on the removal of `:requires_new`.

Specs failed due to `unknown keyword: :requires_new`. From what I can
tell, while that is a valid option for a model's transaction methods, it
isn't for the underlying connection.

My inspection says that `conn.begin_transaction` is sourced from these
two locations (version 5.2 and 6.1):

https://github.com/rails/rails/blob/5-2-stable/activerecord/lib/active_record/connection_adapters/abstract/transaction.rb#L196
https://github.com/rails/rails/blob/6-1-stable/activerecord/lib/active_record/connection_adapters/abstract/transaction.rb#L233

In 5, the method accepted an options hash, but it no longer does in 6.
Reading that code back up through the chain of RealTransaction,
SavepointTransaction, and Transaction I don't see where it every uses
`requires_new`.

Specifically, it seems to only use the `:joinable` option here:

https://github.com/rails/rails/blob/5-2-stable/activerecord/lib/active_record/connection_adapters/abstract/transaction.rb#L97

But I could be entirely wrong about this so would love some confirmation
:-)